Basic implementation of vector.
Uses malloc/realloc for pod type.
For classes uses placement new to insert new copies of the data.

In the test package, renamed macros.h to common.h

New memory file with uninitialized_copy and uninitialized_fill.

Disabled the host target since it is used only on linux to run
the valgrind tool and that I don't want to spend time making it
compile on mac, cygwin ....