Use copy constructor to push new non-pod element in the vector.
In the push_back method:
Previously we used the assignment operator for both pod and non pod types.

This was broken for non pod since the memory at the end of the vector where
we tried to assign to was not initialized (lhs cw not a valid instance).
Instead we use placement new to copy the element to the new location.

In the reserve method:
Fix line length.
Do not try to copy from the old memory area to the new one if the
area is non existent (mBegin is NULL).