Added iterator based vector constructor.
The new range based constructors degrades to one reserve
 and one memmove call for pod types referenced using random
access iterators.

There is an extra step to disambiguate between range and repeat
initializers. Basically for some type combination, the range
form can be called when the repeat form was the one targetted.
We check if the type is an integer to find out which one was called.