Basic streambuf implementation.
Needed to finish the ostream. We are not expected to use this class a
lot since we are going to rely on the stdio implementation for all the
work.
Still needed to make the compiler happy but did not spent too much effort on it.
For instance only output buffer operations are supported.